requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.2
pydantic-settings==2.1.0
websockets==12.0

# AI/LLM
openai==1.6.1
google-cloud-aiplatform==1.38.0
vertexai==0.0.1

# OCR
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.1.0

# Database
sqlalchemy==2.0.23
aiosqlite==0.19.0

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
tenacity==8.2.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
